+++ lib/freydis.rb
@@ -9,6 +9,7 @@ require_relative 'freydis/cryptsetup'
 require_relative 'freydis/rsync'
 require_relative 'freydis/error'
 require_relative 'freydis/guard'
+require_relative 'freydis/secrets'
 
 module Freydis
   CONFIG = Config.new